# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitSphere API",
        "description": "API Documentation for FitSphere",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "member",
                                        "enum": [
                                            "member",
                                            "trainer",
                                            "gym"
                                        ]
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1 (555) 000-0000"
                                    },
                                    "dob": {
                                        "description": "Required for member and trainer",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "fitness_goal": {
                                        "description": "Required for member",
                                        "type": "string",
                                        "example": "Lose weight"
                                    },
                                    "weight": {
                                        "description": "Required for member",
                                        "type": "number",
                                        "example": 70
                                    },
                                    "height": {
                                        "description": "Required for member",
                                        "type": "number",
                                        "example": 175
                                    },
                                    "experience_years": {
                                        "description": "Required for trainer",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "specialization": {
                                        "description": "Required for trainer",
                                        "type": "string",
                                        "example": "Yoga"
                                    },
                                    "price_per_month": {
                                        "description": "Required for trainer",
                                        "type": "number",
                                        "example": 50
                                    },
                                    "bio": {
                                        "description": "Required for trainer",
                                        "type": "string",
                                        "example": "Experienced trainer..."
                                    },
                                    "manager_name": {
                                        "description": "Required for gym",
                                        "type": "string",
                                        "example": "Manager Name"
                                    },
                                    "manager_email": {
                                        "description": "Required for gym",
                                        "type": "string",
                                        "format": "email",
                                        "example": "manager@gym.com"
                                    },
                                    "city": {
                                        "description": "Required for gym",
                                        "type": "string",
                                        "example": "Cairo"
                                    },
                                    "street_name": {
                                        "description": "Required for gym",
                                        "type": "string",
                                        "example": "Tahrir Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/videos": {
            "get": {
                "tags": [
                    "Content - Videos"
                ],
                "summary": "List videos with advanced filtering",
                "operationId": "319b58b9d1ed853fc9d8f9bc1afd2f4f",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category name (e.g. Workout, Recovery)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "muscle",
                        "in": "query",
                        "description": "Filter by muscle name (e.g. Full Body, Chest)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by video title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/videos/{id}": {
            "get": {
                "tags": [
                    "Content - Videos"
                ],
                "summary": "Get single video and related videos",
                "operationId": "10008a4246276b57ee92bc37df44b951",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}